Question_service/Refactor controller and router structure
